+++ packages/codec/lib/abi/index.ts
@@ -1,4 +1,4 @@
-import * as Allocate from "@truffle/codec/allocate/abi";
+import * as Allocate from "./allocate";
 export { Allocate };
 
 import * as Encode from "@truffle/codec/encode/abi";
